+++ src/deps.ts
@@ -1,4 +1,4 @@
-export * as denosass from "./wasm/grass.js";
+export * as denosass from "./wasm/grass.deno.js";
 export * from "https://deno.land/std@0.119.0/testing/asserts.ts";
 export {
   emptyDirSync,
@@ -6,3 +6,4 @@ export {
   walkSync,
 } from "https://deno.land/std@0.125.0/fs/mod.ts";
 export * as path from "https://deno.land/std@0.125.0/path/mod.ts";
+
